Implement options parsing
This is a huge commit as I basically replicated what I did in jrequester but adapted it to typescript.

This commit also contains a whole bunch of tests about options. Both default and custom option prefixes should be covered.

An implementation of the OptionDef class was added but no tests on that have been done as I don't think it will be final. This class will allow the ability to define behaviors for the actual, defined options.

The implementation of options accepting content or not was not implemented yet.

More classes were exposed via the index path, since these will be used more extensively.