Add lazy option parsing possibility
A new method on the VParsedCommand class now allows to set the option definitions after the message has been parsed, allowing to retrieve the parsed name and content to define further flows if required.

This new method also work to update the definitions of an already existing parsed options.

A new parsing function was added to the VCommandParser to actually parse the message lazily.

New tests were added, based on previous tests, to confirm the behavior of parsing options in the future.